Список удаления сделан в виде плиток.
https://www.evernote.com/shard/s132/nl/14501366/a5aa12d0-8033-471f-be9d-4675825e5622

git-svn-id: https://rationalcity.com:8443/svn/Morozov/my/timekeeper/win%20phone%208@2768 d504a619-eb68-f843-b33a-476b659037dc

+++ TimeKeeper/DeleteCategoryPage.xaml.cs
@@ -30,10 +30,14 @@ namespace TimeKeeper
             var item = (Category)CategoryList.SelectedItem;
             if (item != null)
             {
-                var name = item.Name;
-                if (MessageBox.Show(string.Format(AppResources.DeleteCategoryMessage, name), AppResources.DeleteCategoryCaption, MessageBoxButton.OKCancel) == MessageBoxResult.OK)
+                //  Searches the list to protect from duplicate question.
+                if (MainPage.Data.Any.IndexOf(item) != -1)
                 {
-                    MainPage.Data.DeleteCategory(item);
+                    var name = item.Name;
+                    if (MessageBox.Show(string.Format(AppResources.DeleteCategoryMessage, name), AppResources.DeleteCategoryCaption, MessageBoxButton.OKCancel) == MessageBoxResult.OK)
+                    {
+                        MainPage.Data.DeleteCategory(item);
+                    }
                 }
             }
         }
